refactor/utils파일 분리
