+++ docs/index.html
@@ -6,8 +6,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-<script defer src="bundle.js"></script><link href="main.css" rel="stylesheet"></head>
-
+    <script defer src="bundle.js"></script><link href="main.css" rel="stylesheet"></head>
 <body>
     <div id="wrapper">
         <div id="document" class="document">
