+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "webpack-dev-server --config webpack.dev.js",
-    "build": "webpack",
+    "build": "webpack --config webpack.prod.js",
     "dev:": "webpack-dev-server --config webpack.dev.js",
     "prod": "webpack-dev-server --config webpack.prod.js"
   },
