+++ src/components/component.ts
@@ -1,4 +1,10 @@
-import { utils } from '../ts/utils'
+import { draggable } from '../utils/draggable';
+import { elements } from '../utils/attachTo';
+import { consoleUtil } from '../utils/consoleUtil';
+import { attachTo } from '../utils/attachTo';
+import { detectOverlap } from '../utils/detectOverlap';
+import { throttle } from '../utils/throttle';
+import { dragInit } from '../utils/dragaInit';
 
 export class BaseComponent<T extends HTMLElement>  {
 
@@ -6,56 +12,34 @@ export class BaseComponent<T extends HTMLElement>  {
     private parent: HTMLElement
     public isSelected: boolean;
     private mouseMove
-
+S
     constructor(drawOptions) {
 
         // container의 x, y 위치에 element를 그린다.
         const template = document.createElement('template');
         template.innerHTML = drawOptions.htmlString;
 
-        this.parent = drawOptions.container
+        //this.parent = drawOptions.container
+        this.parent = document.body
 
         this.element = template.content.firstElementChild! as T;
         this.element.style.position = 'absolute';
         this.element.style.transform = 'translate(-50%, -50%)'
         this.element.style.left = `${drawOptions.x}px`;
         this.element.style.top = `${drawOptions.y}px`;
-        //this.element.className = 'drag_component';
-
+ 
         this.element.style.backgroundColor = 'rgba(255,0,0,0.2)';
         this.element.style.opacity = '1';
 
-        utils.attachTo(drawOptions, this.element);
-
-        this.mouseMove = () => { utils.throttle(this.detectOverlap(this.element, this.parent)) }
+        attachTo(drawOptions, this.element);
+        elements.push(this.element)
+        
+      
+        this.mouseMove = () => { throttle(detectOverlap(this.element, elements, this.parent)) }
 
         // 드래그 on
-        utils.draggable(this.element, this.parent)
-        utils.dragInit(this.element, this.mouseMove, this.parent);
+        draggable(this.element, this.parent, elements )
+        dragInit(this.element, this.mouseMove, this.parent, elements);
     }
 
-    private detectOverlap(element: T, container: HTMLElement) {
-        // 위의 문법으로 수정
-        const selectedElement = this.element.getBoundingClientRect();
-
-        let childArr = utils.elements;
-        for (const _element of childArr) {
-
-            const elementRect = _element.getBoundingClientRect();
-
-            if (
-                selectedElement.x < elementRect.x + elementRect.width &&
-                selectedElement.x + selectedElement.width > elementRect.x &&
-                selectedElement.y < elementRect.y + elementRect.height &&
-                selectedElement.height + selectedElement.y > elementRect.y
-            ) {
-                _element.style.border = '2px solid blue'
-            } else {
-                _element.style.border = 'none'
-            }
-            //_element.style.border = 'none'
-        }
-
-        this.element.style.border = '2px solid red'
-    }
 }
