+++ src/utils/getPosition.ts
@@ -1,0 +1,4 @@
+export const getPosition = (event, pos) => {
+    pos.x = event.pageX
+    pos.y = event.pageY
+}
